+++ src/main/gen/mensury/edits/melang/psi/MelangEntry.java
@@ -1,0 +1,8 @@
+// This is a generated file. Not intended for manual editing.
+package mensury.edits.melang.psi;
+
+import com.intellij.psi.PsiElement;
+
+public interface MelangEntry extends PsiElement {
+
+}