+++ src/main/gen/mensury/edits/melang/psi/MelangVisitor.java
@@ -1,0 +1,18 @@
+// This is a generated file. Not intended for manual editing.
+package mensury.edits.melang.psi;
+
+import com.intellij.psi.PsiElement;
+import com.intellij.psi.PsiElementVisitor;
+import org.jetbrains.annotations.NotNull;
+
+public class MelangVisitor extends PsiElementVisitor {
+
+  public void visitEntry(@NotNull MelangEntry o) {
+    visitPsiElement(o);
+  }
+
+  public void visitPsiElement(@NotNull PsiElement o) {
+    visitElement(o);
+  }
+
+}