+++ src/main/java/cruelteam/plugins/Icons.java
@@ -1,9 +1,0 @@
-package cruelteam.plugins;
-
-import com.intellij.openapi.util.IconLoader;
-
-import javax.swing.*;
-
-public class Icons {
-    public static final Icon MENSURY = IconLoader.getIcon("/mensury/icon.png", Icons.class);
-}