+++ src/main/java/cruelteam/plugins/mnlang/MnlangLexerAdapter.java
@@ -1,9 +1,0 @@
-package cruelteam.plugins.mnlang;
-
-import com.intellij.lexer.FlexAdapter;
-
-public class MnlangLexerAdapter extends FlexAdapter {
-    public MnlangLexerAdapter() {
-        super(new MnlangLexer(null));
-    }
-}