+++ src/main/java/cruelteam/plugins/mnlang/psi/MnlangTokenSets.java
@@ -1,8 +1,0 @@
-package cruelteam.plugins.mnlang.psi;
-
-import com.intellij.psi.tree.TokenSet;
-
-public interface MnlangTokenSets {
-//    TokenSet IDENTIFIERS = TokenSet.create(MnlangTypes.KEY);
-    TokenSet COMMENTS = TokenSet.create(MnlangTypes.COMMENT);
-}