+++ src/main/java/mensury/edits/Icons.java
@@ -1,0 +1,9 @@
+package mensury.edits;
+
+import com.intellij.openapi.util.IconLoader;
+
+import javax.swing.*;
+
+public class Icons {
+    public static final Icon MENSURY = IconLoader.getIcon("/mensury/melang-icon.png", Icons.class);
+}