+++ src/main/java/mensury/edits/melang/MelangLexerAdapter.java
@@ -1,0 +1,9 @@
+package mensury.edits.melang;
+
+import com.intellij.lexer.FlexAdapter;
+
+public class MelangLexerAdapter extends FlexAdapter {
+    public MelangLexerAdapter() {
+        super(new MelangLexer(null));
+    }
+}