+++ src/main/java/mensury/edits/melang/MelangSyntaxHighlighterFactory.java
@@ -1,0 +1,13 @@
+package mensury.edits.melang;
+
+import com.intellij.openapi.fileTypes.SyntaxHighlighter;
+import com.intellij.openapi.fileTypes.SyntaxHighlighterFactory;
+import com.intellij.openapi.project.Project;
+import com.intellij.openapi.vfs.VirtualFile;
+import org.jetbrains.annotations.NotNull;
+
+public class MelangSyntaxHighlighterFactory extends SyntaxHighlighterFactory {
+    @NotNull @Override public SyntaxHighlighter getSyntaxHighlighter(Project project, VirtualFile virtualFile) {
+        return new MelangSyntaxHighlighter();
+    }
+}