+++ src/main/java/mensury/edits/melang/psi/MelangTokenSets.java
@@ -1,0 +1,8 @@
+package mensury.edits.melang.psi;
+
+import com.intellij.psi.tree.TokenSet;
+
+public interface MelangTokenSets {
+//    TokenSet IDENTIFIERS = TokenSet.create(MelangTypes.KEY);
+    TokenSet COMMENTS = TokenSet.create(MelangTypes.COMMENT);
+}