+++ CHANGELOG.md
@@ -4,6 +4,13 @@
 
 ## [Unreleased]
 
+## [1.0.2]
+- Renamed everything called as "Mnlang" to "Melang"
+- Renamed group to mensury.edits
+- Decreased minimum supported IntelliJ IDEA version to 2021.1
+- Default extension was set to .melang, but .mnlang will also be supported
+- Removed unused tests
+
 ## [1.0.1]
 - Adjusted currently supporting IntelliJ IDEA versions
 - Fix renaming plugin to Mensury Edits